Add support for UTF8 senders/recipients
If the sender or receiver has utf8 characters in their address, then the `MAIL` command must end with `SMTPUTF8`. (if the server supports it) Otherwise, the server will complain when it receives 8 bit data.

Since this library has no knowledge of the recipients when it sends the `MAIL` command (which is fair since this is a "low level" library anyway) we have no way of knowing whether or not we even need to enable `SMTPUTF8` extensions. So that responsibility can fall on the people who end up using this thing.